requirements.txt
Flask==2.3.2
Flask_JWT_Extended==4.5.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.5
PyJWT==2.8.0
PyJWT==2.8.0
PyMySQL==1.1.0
